# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEO APP API",
    "version": "1.0.0",
    "description": "REST API ของร้าน NEO APP (ขายแอปพลิเคชัน + กระเป๋าเงิน)\n\n## การยืนยันตัวตน\n\nทุก endpoint ต้องผ่านการยืนยันตัวตน ไม่มีเส้นไหนเปิดสาธารณะ รองรับ 2 วิธี:\n\n1. **Cookie session (เบราว์เซอร์)** — ล็อกอินที่ `/login` แล้ว Supabase จะเก็บ JWT ไว้ใน cookie `sb-<project-ref>-auth-token`\n2. **Bearer token (Postman / curl / script)** — ส่งเฮดเดอร์ `Authorization: Bearer <access_token>`\n\nขอ access token:\n\n```bash\ncurl -X POST \"$NEXT_PUBLIC_SUPABASE_URL/auth/v1/token?grant_type=password\" \\\n  -H \"apikey: $NEXT_PUBLIC_SUPABASE_PUBLISHABLE_KEY\" \\\n  -H \"content-type: application/json\" \\\n  -d '{\"email\":\"you@example.com\",\"password\":\"...\"}'\n```\n\n## สิทธิ์\n\nข้อมูลถูกคุมด้วย Row Level Security ของ Postgres: ลูกค้าเห็นเฉพาะคำสั่งซื้อ/กระเป๋าเงินของตัวเอง\nส่วนการเพิ่ม-แก้-ลบสินค้า อัปโหลดรูป และแก้ตั้งค่าร้าน ทำได้เฉพาะบัญชีที่มี `app_metadata.role = \"admin\"`\n\n## เอกสารนี้\n\nหน้า `/docs` และไฟล์ `/openapi.json` เปิดให้เฉพาะแอดมิน คนอื่นจะได้ 404",
    "contact": {
      "name": "NEO APP"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/topups": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "ประวัติการเติมเงินที่ผ่านการตรวจสลิปแล้ว",
        "operationId": "listTopups",
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Topup"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "เติมเงินด้วยสลิปโอน (ตรวจกับธนาคารผ่าน RDCW)",
        "description": "ขั้นตอนตรวจก่อนเติมเงินเข้าให้:\n\n1. ส่งสลิปไปตรวจกับ RDCW (https://slip.rdcw.co.th/docs)\n2. ยอดในสลิปต้องตรงกับ `amount` ที่ผู้ใช้กรอก\n3. ผู้รับในสลิปต้องตรงกับบัญชีร้านที่ตั้งไว้ใน `/api/settings`\n4. สลิปต้องไม่เก่ากว่าที่ตั้งไว้ และเลขอ้างอิงต้องไม่เคยถูกใช้มาก่อน (unique ทั้งระบบ)",
        "operationId": "createTopup",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "ยอดที่โอน (บาท) ต้องตรงกับสลิป"
                  },
                  "slip": {
                    "type": "string",
                    "format": "binary",
                    "description": "รูปสลิป JPG/PNG"
                  },
                  "payload": {
                    "type": "string",
                    "description": "ข้อความ QR จากสลิป (ใช้แทนรูปได้)"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "payload"
                ],
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "payload": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "เติมเงินสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "amount": {
                          "type": "number"
                        },
                        "transRef": {
                          "type": "string"
                        },
                        "balance": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "สลิปไม่ผ่านการตรวจ (ยอดไม่ตรง, ไม่ใช่บัญชีร้าน, สลิปเก่าเกินกำหนด)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "สลิปนี้ถูกใช้เติมเงินไปแล้ว",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "ติดต่อระบบตรวจสลิปไม่ได้ หรือโควตา RDCW หมด",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "ร้านยังไม่ได้ตั้งค่าบัญชีรับเงิน",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/uploads": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "summary": "อัปโหลดรูปสินค้าไป Supabase Storage (แอดมินเท่านั้น)",
        "operationId": "uploadImage",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "อัปโหลดสำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "path": {
                          "type": "string"
                        },
                        "url": {
                          "type": "string",
                          "format": "uri"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "description": "ไม่ใช่แอดมิน",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Cookie session ของ Supabase (ชื่อจริงคือ `sb-<project-ref>-auth-token`) เบราว์เซอร์ส่งให้อัตโนมัติหลังล็อกอิน"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token (JWT) สำหรับ API client ที่ไม่มี cookie jar"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "ยังไม่ได้ล็อกอิน หรือ token หมดอายุ/ไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": "unauthorized",
              "message": "กรุณาเข้าสู่ระบบก่อนเรียกใช้งาน API"
            }
          }
        }
      },
      "BadRequest": {
        "description": "ข้อมูลที่ส่งมาไม่ครบหรือไม่ถูกต้อง",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "ข้อผิดพลาดฝั่งเซิร์ฟเวอร์",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Topup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number"
          },
          "transRef": {
            "type": "string",
            "description": "เลขอ้างอิงจากธนาคาร ใช้ซ้ำไม่ได้"
          },
          "senderName": {
            "type": "string"
          },
          "receiverName": {
            "type": "string"
          },
          "transferredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}